Add in the ability to configure the Date Transformers display 'depth'.
This means that rather than just taking the first match for a date-time technically 1 year, 3 months, 2 days in the past and showing "One year", this will now take the input of RelativeTransformDepth, which defaults to current behavior, but can then return one, two, or three deep.

This change only includes the English file, and I am not 100% convinced this will work in all languages depending on the formatting that is required. As the formatter iterates over the possible NSDateComponents, each component is now collected as a date/value pair first. After these are collected to the desired depth, the proper template is then rendered. This is different than the existing implementation, which only had the one possible template.